# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: CCAT Practice App — Gateway API
  version: "1.0.0-phase0"
  description: >
    Normative /v1 API for the Concept Mastery CCAT Practice App. The Gateway is the ONLY
    application-data boundary exposed to the student app (Blueprint v10.0 §1.1, §33).
    Direct student Supabase/PostgREST access is prohibited (§41). All sensitive mutations
    use idempotency keys; all mutable Admin resources use expected-version/ETag concurrency
    (§22.4, §32.1). This document is Phase 0 and covers the endpoint surface of §32; request
    bodies for Admin write operations are represented by shared envelopes and will be
    expanded per-domain in Phase 1.
  contact:
    name: Concept Mastery
    email: admin@conceptmastery.ca
servers:
  - url: https://api.conceptmastery.ca/v1
    description: Production
  - url: https://staging-api.conceptmastery.ca/v1
    description: Staging (no real child PII, §38.1)
...
security:
  - studentSession: []
paths:
...
  /admin/accounts:
    post:
      tags: [admin-accounts]
      operationId: adminCreateAdmin
      security: [ { adminSession: [] } ]
      summary: Provision Admin with temp password (Super-Admin, §22.2)
      responses:
        "201": { description: Admin created; temp password shown once }
        "403": { $ref: "#/components/responses/Error" }
  /admin/accounts/{id}:
    patch:
      tags: [admin-accounts]
      operationId: adminUpdateAdmin
      security: [ { adminSession: [] } ]
      summary: Enable/disable, set permissions (Super-Admin, §28.1). Protects last SA (§28.2).
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - { $ref: "#/components/parameters/IfMatch" }
      responses:
        "200": { description: Updated }
        "403": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Conflict" }
components:
  securitySchemes:
    studentSession:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Student application session token, bound to the sole enrolled device (§5.4).
    adminSession:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Admin identity token (email/password + mandatory MFA, §22.1). Permissions loaded server-side.
  parameters:
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema: { type: string }
      description: Expected version/ETag for mutable Admin resources (§22.4). Missing/stale => 409.
...
  responses:
    Error:
      description: Structured error envelope (§32.1)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: Optimistic-concurrency conflict with comparison (§22.4)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ConflictError" }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, description: Machine code, e.g. ACTIVE_SESSION_EXISTS }
            message: { type: string }
            request_id: { type: string }
            details: { type: object, additionalProperties: true }
    ConflictError:
      allOf:
        - { $ref: "#/components/schemas/Error" }
        - type: object
          properties:
            your_version: { type: integer }
            current_version: { type: integer }
            current: { type: object, additionalProperties: true, description: Current server representation for the diff UI }
...